remove core dependency from api project
