+++ Sockseek.Api/Client/SockseekApiJsonContext.cs
@@ -128,12 +128,12 @@ namespace Sockseek.Api;
 [JsonSerializable(typeof(JoinRoomRequestDto))]
 [JsonSerializable(typeof(AddRoomMemberRequestDto))]
 [JsonSerializable(typeof(MarkNotificationsReadRequestDto))]
-[JsonSerializable(typeof(Sockseek.Core.Chat.ChatTargetKind))]
-[JsonSerializable(typeof(Sockseek.Core.Chat.ChatMessageDirection))]
-[JsonSerializable(typeof(Sockseek.Core.Chat.ChatMessageState))]
-[JsonSerializable(typeof(Sockseek.Core.Chat.ChatRoomKind))]
-[JsonSerializable(typeof(Sockseek.Core.Chat.ChatRoomJoinPhase))]
-[JsonSerializable(typeof(Sockseek.Core.Chat.UserNotificationKind))]
+[JsonSerializable(typeof(ServerChatTargetKind))]
+[JsonSerializable(typeof(ServerChatMessageDirection))]
+[JsonSerializable(typeof(ServerChatMessageState))]
+[JsonSerializable(typeof(ServerChatRoomKind))]
+[JsonSerializable(typeof(ServerChatRoomJoinPhase))]
+[JsonSerializable(typeof(ServerUserNotificationKind))]
 [JsonSerializable(typeof(StartShareScanResult))]
 [JsonSerializable(typeof(StartShareScanResponseDto))]
 [JsonSerializable(typeof(TransferQueueEstimateDto))]
@@ -144,8 +144,7 @@ namespace Sockseek.Api;
 [JsonSerializable(typeof(TransferTerminalOutcome))]
 [JsonSerializable(typeof(TransferFailureReason))]
 [JsonSerializable(typeof(TransferCancellationSource))]
-[JsonSerializable(typeof(Sockseek.Core.Sharing.ShareBrowseStatus))]
-[JsonSerializable(typeof(Sockseek.Core.Sharing.ShareScanPhase))]
+[JsonSerializable(typeof(ServerShareScanPhase))]
 [JsonSerializable(typeof(JobDisplayFieldsDto))]
 [JsonSerializable(typeof(JobLifecycleFieldsDto))]
 [JsonSerializable(typeof(JobDiscoveryFieldsDto))]
@@ -194,7 +193,14 @@ namespace Sockseek.Api;
 [JsonSerializable(typeof(ResolveEffectiveSettingsResponseDto))]
 [JsonSerializable(typeof(EffectiveDownloadSettingsDto))]
 [JsonSerializable(typeof(DownloadBehaviorPolicyDto))]
-[JsonSerializable(typeof(Sockseek.Core.DownloadBehavior))]
+[JsonSerializable(typeof(ServerDownloadBehavior))]
+[JsonSerializable(typeof(ServerExtractionMode))]
+[JsonSerializable(typeof(ServerSearchDefaultProjectionKind))]
+[JsonSerializable(typeof(ServerSearchViewProjectionKind))]
+[JsonSerializable(typeof(ServerSearchResultVisibility))]
+[JsonSerializable(typeof(ServerSearchPreferenceTier))]
+[JsonSerializable(typeof(ServerSearchPreferenceCondition))]
+[JsonSerializable(typeof(ServerSearchSettingsBaselineKind))]
 
 [JsonSerializable(typeof(JobProvenanceDto))]
 [JsonSerializable(typeof(SourceMutationDto))]
@@ -257,7 +263,11 @@ namespace Sockseek.Api;
 [JsonSerializable(typeof(DownloadSettingsPatchDto))]
 [JsonSerializable(typeof(OutputSettingsPatchDto))]
 [JsonSerializable(typeof(IncompleteAlbumActionSettingsPatchDto))]
-[JsonSerializable(typeof(Sockseek.Core.IncompleteAlbumActionKind))]
+[JsonSerializable(typeof(ServerPrintOption))]
+[JsonSerializable(typeof(ServerIncompleteAlbumActionKind))]
+[JsonSerializable(typeof(ServerAlbumArtOption))]
+[JsonSerializable(typeof(ServerSkipMode))]
+[JsonSerializable(typeof(ServerInputType))]
 [JsonSerializable(typeof(SearchSettingsPatchDto))]
 [JsonSerializable(typeof(FileConditionsPatchDto))]
 [JsonSerializable(typeof(FolderConditionsPatchDto))]
