+++ Sockseek.Api/Sockseek.Api.csproj
@@ -11,7 +11,6 @@
   <ItemGroup>
     <PackageReference Include="Microsoft.AspNetCore.SignalR.Client" Version="10.0.9" />
     <PackageReference Include="Microsoft.AspNetCore.SignalR.Protocols.Json" Version="10.0.9" />
-    <ProjectReference Include="..\Sockseek.Core\Sockseek.Core.csproj" />
   </ItemGroup>
 
 </Project>
