+++ Sockseek.Server.Tests/ChatLiveStateTests.cs
@@ -2,6 +2,7 @@ using System.Text;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Sockseek.Api;
 using Sockseek.Core.Chat;
+using Sockseek.Server;
 
 namespace Tests.Server;
 
@@ -15,7 +16,7 @@ public sealed class ChatLiveStateTests
         Guid epoch = Guid.NewGuid();
         Guid conversationId = Guid.NewGuid();
         StateStreamScopeDto scope = StateStreamScopeDto.ChatConversation(conversationId);
-        ChatMessageDto first = Message(conversationId, 1, "first", ChatMessageState.Received);
+        ChatMessageDto first = Message(conversationId, 1, "first", ServerChatMessageState.Received);
         var conversation = Conversation(conversationId, 1, first);
         store.ApplySnapshot(new StateSnapshotDto(
             scope,
@@ -24,9 +25,9 @@ public sealed class ChatLiveStateTests
             null,
             [], [], [], [],
             new ChatTargetSnapshotDto(
-                ChatTargetKind.Direct, conversationId, conversation, null, [first], false)));
+                ServerChatTargetKind.Direct, conversationId, conversation, null, [first], false)));
 
-        ChatMessageDto second = Message(conversationId, 2, "second", ChatMessageState.Received);
+        ChatMessageDto second = Message(conversationId, 2, "second", ServerChatMessageState.Received);
         var update = store.Apply(new StateUpdateBatchDto(
             scope,
             epoch,
@@ -38,7 +39,7 @@ public sealed class ChatLiveStateTests
                 ChatTargets =
                 [
                     new ChatTargetDeltaDto(
-                        ChatTargetKind.Direct,
+                        ServerChatTargetKind.Direct,
                         conversationId,
                         Conversation(conversationId, 2, second),
                         Messages: [second])
@@ -61,7 +62,7 @@ public sealed class ChatLiveStateTests
         Guid epoch = Guid.NewGuid();
         Guid conversationId = Guid.NewGuid();
         StateStreamScopeDto scope = StateStreamScopeDto.ChatConversation(conversationId);
-        ChatMessageDto first = Message(conversationId, 1, "first", ChatMessageState.Received);
+        ChatMessageDto first = Message(conversationId, 1, "first", ServerChatMessageState.Received);
         store.ApplySnapshot(new StateSnapshotDto(
             scope,
             new StateStreamPositionDto(epoch, 0),
@@ -69,7 +70,7 @@ public sealed class ChatLiveStateTests
             null,
             [], [], [], [],
             new ChatTargetSnapshotDto(
-                ChatTargetKind.Direct,
+                ServerChatTargetKind.Direct,
                 conversationId,
                 Conversation(conversationId, 1, first),
                 null,
@@ -86,7 +87,7 @@ public sealed class ChatLiveStateTests
             {
                 ChatTargets =
                 [
-                    new ChatTargetDeltaDto(ChatTargetKind.Direct, Guid.NewGuid())
+                    new ChatTargetDeltaDto(ServerChatTargetKind.Direct, Guid.NewGuid())
                 ],
             },
             [])));
@@ -104,10 +105,10 @@ public sealed class ChatLiveStateTests
             null,
             [], [], [], []));
         Guid conversationId = Guid.NewGuid();
-        ChatMessageDto message = Message(conversationId, 1, "hello", ChatMessageState.Received);
+        ChatMessageDto message = Message(conversationId, 1, "hello", ServerChatMessageState.Received);
         var notification = new UserNotificationDto(
-            Guid.NewGuid(), 1, UserNotificationKind.PrivateMessage,
-            DateTimeOffset.UtcNow, null, "Alice", ChatTargetKind.Direct,
+            Guid.NewGuid(), 1, ServerUserNotificationKind.PrivateMessage,
+            DateTimeOffset.UtcNow, null, "Alice", ServerChatTargetKind.Direct,
             conversationId, "Alice", message.MessageId, "hello",
             $"/api/chat/conversations/{conversationId:D}");
 
@@ -152,7 +153,7 @@ public sealed class ChatLiveStateTests
             DateTimeOffset.UtcNow,
             StateDeltaDto.Empty with
             {
-                ChatTargets = [new ChatTargetDeltaDto(ChatTargetKind.Direct, conversationId)],
+                ChatTargets = [new ChatTargetDeltaDto(ServerChatTargetKind.Direct, conversationId)],
             },
             [])));
 
@@ -193,7 +194,7 @@ public sealed class ChatLiveStateTests
         Guid epoch = Guid.NewGuid();
         Guid conversationId = Guid.NewGuid();
         StateStreamScopeDto scope = StateStreamScopeDto.ChatConversation(conversationId);
-        ChatMessageDto message = Message(conversationId, 1, "hello", ChatMessageState.Received);
+        ChatMessageDto message = Message(conversationId, 1, "hello", ServerChatMessageState.Received);
         store.ApplySnapshot(new StateSnapshotDto(
             scope,
             new StateStreamPositionDto(epoch, 1),
@@ -201,7 +202,7 @@ public sealed class ChatLiveStateTests
             null,
             [], [], [], [],
             new ChatTargetSnapshotDto(
-                ChatTargetKind.Direct,
+                ServerChatTargetKind.Direct,
                 conversationId,
                 Conversation(conversationId, 1, message),
                 null,
@@ -221,7 +222,7 @@ public sealed class ChatLiveStateTests
         Guid epoch = Guid.NewGuid();
         Guid conversationId = Guid.NewGuid();
         StateStreamScopeDto scope = StateStreamScopeDto.ChatConversation(conversationId);
-        ChatMessageDto first = Message(conversationId, 1, "first", ChatMessageState.Received);
+        ChatMessageDto first = Message(conversationId, 1, "first", ServerChatMessageState.Received);
         store.ApplySnapshot(new StateSnapshotDto(
             scope,
             new StateStreamPositionDto(epoch, 1),
@@ -229,7 +230,7 @@ public sealed class ChatLiveStateTests
             null,
             [], [], [], [],
             new ChatTargetSnapshotDto(
-                ChatTargetKind.Direct,
+                ServerChatTargetKind.Direct,
                 conversationId,
                 Conversation(conversationId, 1, first),
                 null,
@@ -247,7 +248,7 @@ public sealed class ChatLiveStateTests
                 ChatTargets =
                 [
                     new ChatTargetDeltaDto(
-                        ChatTargetKind.Direct,
+                        ServerChatTargetKind.Direct,
                         conversationId,
                         Conversation(conversationId, 2, first) with { LastMessage = null },
                         Messages: [],
@@ -269,13 +270,13 @@ public sealed class ChatLiveStateTests
         Guid conversationId = Guid.NewGuid();
         Guid messageId = Guid.NewGuid();
         var message = new ChatMessageRecord(
-            messageId, 1, "local", ChatTargetKind.Direct, conversationId,
-            "alice", "Alice", "alice", "Alice", ChatMessageDirection.Incoming,
+            messageId, 1, "local", Sockseek.Core.Chat.ChatTargetKind.Direct, conversationId,
+            "alice", "Alice", "alice", "Alice", Sockseek.Core.Chat.ChatMessageDirection.Incoming,
             "first\n" + new string('x', ChatLimits.NotificationPreviewCharacters + 20),
-            DateTimeOffset.UtcNow, DateTimeOffset.UtcNow, ChatMessageState.Received,
+            DateTimeOffset.UtcNow, DateTimeOffset.UtcNow, Sockseek.Core.Chat.ChatMessageState.Received,
             null, 1, DateTimeOffset.UtcNow);
         var record = new UserNotificationRecord(
-            Guid.NewGuid(), 1, "local", UserNotificationKind.PrivateMessage,
+            Guid.NewGuid(), 1, "local", Sockseek.Core.Chat.UserNotificationKind.PrivateMessage,
             messageId, DateTimeOffset.UtcNow, null, message);
 
         UserNotificationDto projected = ChatDtoMapper.ToDto(record);
@@ -287,10 +288,10 @@ public sealed class ChatLiveStateTests
     }
 
     private static ChatMessageDto Message(
-        Guid targetId, long sequence, string text, ChatMessageState state)
+        Guid targetId, long sequence, string text, ServerChatMessageState state)
         => new(
-            Guid.NewGuid(), sequence, ChatTargetKind.Direct, targetId, "Alice",
-            ChatMessageDirection.Incoming, text, DateTimeOffset.UtcNow,
+            Guid.NewGuid(), sequence, ServerChatTargetKind.Direct, targetId, "Alice",
+            ServerChatMessageDirection.Incoming, text, DateTimeOffset.UtcNow,
             DateTimeOffset.UtcNow, state, null);
 
     private static ConversationSummaryDto Conversation(
@@ -301,8 +302,8 @@ public sealed class ChatLiveStateTests
     {
         Guid messageId = Guid.NewGuid();
         return new UserNotificationDto(
-            Guid.NewGuid(), sequence, UserNotificationKind.PrivateMessage,
-            DateTimeOffset.UtcNow, null, "Alice", ChatTargetKind.Direct,
+            Guid.NewGuid(), sequence, ServerUserNotificationKind.PrivateMessage,
+            DateTimeOffset.UtcNow, null, "Alice", ServerChatTargetKind.Direct,
             conversationId, "Alice", messageId, "hello",
             $"/api/chat/conversations/{conversationId:D}");
     }
