+++ webui/DAEMON-AUDIT.md
@@ -97,6 +97,4 @@ The current live activity/state edge is intentionally ephemeral and best-effort 
 
 ### API refactor
 
-`Sockseek.Api` should be split so the public wire contract and reusable remote client no longer depend on `Sockseek.Core`. Today the API project mixes DTOs/protocol types, client code, and server-side mappings back into Core settings, which means a consumer of the API surface can inherit dependencies that belong to the daemon implementation rather than to a remote client. A cleaner structure would put pure DTOs and protocol enums in a small contracts assembly, put `SockseekApiClient`, `SockseekLiveClient`, and `DaemonClientStore` in a lightweight client assembly that depends only on those contracts plus HTTP/SignalR, and keep all Core↔API mapping on the server side.
-
-That refactor is worthwhile regardless of whether the production WebUI uses Blazor or Svelte. It gives the daemon a genuinely reusable, implementation-independent public client boundary, keeps remote CLI concerns clean, prevents accidental leakage of engine dependencies into consumers, and makes the API easier to version and test. If the WebUI is Blazor, it can reference the same browser-safe client directly; if it is Svelte, the split still provides a much clearer canonical contract and reference implementation for the TypeScript client.
+`Sockseek.Api` should become an implementation-independent remote SDK and stop depending on `Sockseek.Core`. Keep its DTOs, protocol enums, JSON metadata, `SockseekApiClient`, `SockseekLiveClient`, and `DaemonClientStore` together for now; give wire-visible enums API-owned definitions, and move Core↔API mapping and Core-dependent validation to one application-side owner shared by the daemon and local CLI. Add an architecture test preventing public API contracts from referencing Core types. A separate contracts/client assembly split is common but not yet necessary here; introduce it only when independent distribution, browser/AOT compatibility, or another concrete consumer requires it.
